+++ NEWS.md
@@ -14,6 +14,12 @@ the dosing including dose amount and route.
   the end of the latest dose administration (#139)
 * `aucint.inf.pred` is `NA` when half-life is not estimable (#450)
 
+## New features
+
+* PKNCA now has a debugging mode to support troubleshooting; it is not intended
+  for production use. Debugging mode can be enabled using
+  `PKNCA.options(debug = TRUE)`.
+
 # PKNCA 0.12.0
 
 ## Breaking changes